% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arusha.R
\docType{data}
\name{arusha_df}
\alias{arusha_df}
\title{Example dataset for the agera5 package}
\format{
An object of class \code{data.frame} with 100 rows and 4 columns.
}
\usage{
arusha_df
}
\description{
100 points in Arusha, Tanzania
The geographic coordinates were generated with the function st_sample from package sf
}
\keyword{datasets}
