% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_agera5}
\alias{download_agera5}
\title{Downloads "Agrometeorological indicators from 1979 to present derived from reanalysis" data using the cdsapi
Data is downloaded from Copernicus Climate Change Service (C3) under de license terms
stated in the license terms documentation:
https://cds.climate.copernicus.eu/api/v2/terms/static/licence-to-use-copernicus-products.pdf
We only provide a programmatic bridge between R code and the cdsapi to facilitate R users the access to
AgeERA5 data.
This code is open source without the kind of warranty.}
\usage{
download_agera5(
  agera5_var,
  agera5_stat = NULL,
  day = NULL,
  month = NULL,
  year,
  path
)
}
\arguments{
\item{day}{Day of the week for the requested data. NULL will download all days from requested month}

\item{month}{Month to be requested. NULL will dowload all the months for the requested year.}

\item{year}{Year to download. Should be between 1979 - 2020}

\item{variable}{Requested variable, e.g. 2m Land Surface Temperature}

\item{statistic}{Requested statistic for the requested variable, e.g.}
}
\description{
Downloads "Agrometeorological indicators from 1979 to present derived from reanalysis" data using the cdsapi
Data is downloaded from Copernicus Climate Change Service (C3) under de license terms
stated in the license terms documentation:
https://cds.climate.copernicus.eu/api/v2/terms/static/licence-to-use-copernicus-products.pdf
We only provide a programmatic bridge between R code and the cdsapi to facilitate R users the access to
AgeERA5 data.
This code is open source without the kind of warranty.
}
