% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agera5_extract.R
\name{extract_period}
\alias{extract_period}
\title{extracts data for a given location and time period}
\usage{
extract_period(
  .start_date,
  .end_date,
  .location_xy,
  .var,
  .statistic,
  .agera5_folder
)
}
\description{
extracts data for a given location and time period
}
