% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agera5_extract.R
\name{extract_single_date}
\alias{extract_single_date}
\title{extracts data for a given location and date}
\usage{
extract_single_date(.date, .location_xy, .var, .statistic)
}
\description{
extracts data for a given location and date
}
