% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prec.R
\name{get_prec}
\alias{get_prec}
\alias{get_prec.data_point}
\alias{get_prec.period}
\alias{get_prec.dataset}
\title{This set of functions extract precipitation data from locally stored AgERA5 data}
\usage{
get_prec.data_point(.date, .location_xy, .agera5_folder)

get_prec.period(.start_date, .end_date, .location_xy, .agera5_folder)

get_prec.dataset(
  .trial_dataset = NULL,
  .start_date = "pdate",
  .end_date = "hdate",
  .lon = "lon",
  .lat = "lat",
  .agera5_folder
)
}
\arguments{
\item{.date}{a date or character representing the date of the point data to be extracted}

\item{.location_xy}{a data.frame or an object to be coerced, with longitude and latitude}

\item{.var}{a character of the variable of interest}

\item{.statistic}{an integer for the statistic to extract, options are:}
}
\description{
This set of functions extract precipitation data from locally stored AgERA5 data
}
\section{Functions}{
\itemize{
\item \code{get_prec.period}: Get precipitation data for one location for a provided time period

\item \code{get_prec.dataset}: Iterates across a data set to extract all required data points
}}

