% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prec.R
\name{get_prec_dp}
\alias{get_prec_dp}
\title{Extract precipitation data for one location and one date}
\usage{
get_prec_dp(.date, .lon, .lat, .agera5_folder)
}
\arguments{
\item{.date}{a date or character representing the date of the data point to be extracted}

\item{.lon}{numeric Longitude for the point of interest}

\item{.lat}{numeric Latitude for the point of interest}

\item{.agera5_folder}{Folder in the local system where the agera5 data is stored.}
}
\description{
Extract precipitation data for one location and one date
}
