% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prec.R
\name{get_prec_ds}
\alias{get_prec_ds}
\title{Extract precipitation data for a dataset}
\usage{
get_prec_ds(
  .trial_dataset = NULL,
  .start_date = "pdate",
  .end_date = "hdate",
  .lon = "lon",
  .lat = "lat",
  .agera5_folder
)
}
\arguments{
\item{.trial_dataset}{data.frame with columns required}

\item{.start_date}{character Name of the column in the dataset}

\item{.end_date}{character Name of the column in the dataset}

\item{.lon}{chracter Name of the column in the dataset}

\item{.lat}{chracter Name of the column in the dataset}

\item{.agera5_folder}{Folder in the local system where the agera5 data is stored.}
}
\value{
a list of numeric vectors with requested precipation values
}
\description{
Extract precipitation data for a dataset
}
