% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prec.R
\name{get_prec_ts}
\alias{get_prec_ts}
\title{Extract precipitation data for one location and one time series}
\usage{
get_prec_ts(.start_date, .end_date, .lon, .lat, .agera5_folder)
}
\arguments{
\item{.start_date}{Date or character to be coerced as Date The starting date for the period to extract}

\item{.end_date}{Date or character to be coerced as Date The end date for the period to extract}

\item{.lon}{numeric Longitude for the point of interest}

\item{.lat}{numeric Latitude for the point of interest}

\item{.agera5_folder}{Folder in the local system where the AgERA5 data is stored.}
}
\value{
numeric vector with precipitation values
}
\description{
Extract precipitation data for one location and one time series
}
