% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rhum.R
\name{get_rhum}
\alias{get_rhum}
\alias{get_rhum.data_point}
\alias{get_rhum.period}
\alias{get_rhum.dataset}
\title{Extract relative humidity values from locally stored AgERA5 data}
\usage{
get_rhum(...)

\method{get_rhum}{data_point}(.date, .lon, .lat, .time, .agera5_folder)

\method{get_rhum}{period}(.start_date, .end_date, .lon, .lat, .time, .agera5_folder)

\method{get_rhum}{dataset}(
  .trial_dataset = NULL,
  .start_date = "pdate",
  .end_date = "hdate",
  .lon = "lon",
  .lat = "lat",
  .time = "time",
  .agera5_folder
)
}
\arguments{
\item{.date}{a date or character representing the date of the point data to be extracted}

\item{.lon}{character Name of the column that holds the longitude}

\item{.lat}{character time of the day from the options available in AgERA5 dataset, see details}

\item{.time}{a data.frame or an object to be coerced, with longitude and latitude}

\item{.agera5_folder}{a character with agera5 data folder location}

\item{.start_date}{character Name of the column that holds the start date of the time period to extract}

\item{.end_date}{character Name of the column that holds the end date of the time period to extract}

\item{.trial_dataset}{data.frame containing the required parameters, usually the trial data points}
}
\description{
Extract relative humidity values from locally stored AgERA5 data
}
\details{
\strong{valid .time values}
\itemize{
\item 06h
\item 09h
\item 12h
\item 15h
\item 18h
}
}
\section{Methods (by class)}{
\itemize{
\item \code{period}: Get temperature data for one location for a provided time period

\item \code{dataset}: Iterates across a data set to extract all required data points
}}

