% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rhum.R
\name{get_rhum_dp}
\alias{get_rhum_dp}
\title{Extract relative humidity data for one location and one date}
\usage{
get_rhum_dp(.date, .lon, .lat, .time, .agera5_folder)
}
\arguments{
\item{.date}{a date or character representing the date of the point data to be extracted}

\item{.lon}{a data.frame or an object to be coerced, with longitude and latitude}

\item{.lat}{a data.frame or an object to be coerced, with longitude and latitude}

\item{.time}{a data.frame or an object to be coerced, with longitude and latitude}

\item{.agera5_folder}{character indicating the folder where agera5 nc files are located}
}
\description{
Extract relative humidity data for one location and one date
}
\details{
\strong{valid .time values}
\itemize{
\item 06h
\item 09h
\item 12h
\item 15h
\item 18h
}
}
