% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rhum.R
\name{get_rhum_ds}
\alias{get_rhum_ds}
\title{Extract relative humidity data for dataset}
\usage{
get_rhum_ds(
  .trial_dataset = NULL,
  .start_date = "pdate",
  .end_date = "hdate",
  .lon = "lon",
  .lat = "lat",
  .time = "time",
  .agera5_folder
)
}
\arguments{
\item{.trial_dataset}{data.frame containing the required parameters, usually the trial data points}

\item{.start_date}{character Name of the column that holds the start date of the time period to extract}

\item{.end_date}{character Name of the column that holds the end date of the time period to extract}

\item{.lon}{character Name of the column that holds the longitude}

\item{.lat}{character time of the day from the options available in AgERA5 dataset, see details}

\item{.agera5_folder}{a character with agera5 data folder location}
}
\description{
Extract relative humidity data for dataset
}
