% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rhum.R
\name{get_rhum_ts}
\alias{get_rhum_ts}
\title{Extract relative humidity data for one location and one time series}
\usage{
get_rhum_ts(.start_date, .end_date, .lon, .lat, .time, .agera5_folder)
}
\description{
Extract relative humidity data for one location and one time series
}
