% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_srf.R
\name{get_srf}
\alias{get_srf}
\alias{get_srf.data_point}
\alias{get_srf.period}
\alias{get_srf.dataset}
\title{This set of functions extract solar radiatio flux data from locally stored AgERA5 data}
\usage{
get_srf.data_point(.date, .lon, .lat, .agera5_folder)

get_srf.period(.start_date, .end_date, .lon, .lat, .agera5_folder)

get_srf.dataset(
  .trial_dataset = NULL,
  .start_date = "pdate",
  .end_date = "hdate",
  .lon = "lon",
  .lat = "lat",
  .agera5_folder
)
}
\arguments{
\item{.date}{a date or character representing the date of the point data to be extracted}

\item{.location_xy}{a data.frame or an object to be coerced, with longitude and latitude}

\item{.var}{a character of the variable of interest}

\item{.statistic}{an integer for the statistic to extract, options are:}
}
\description{
This set of functions extract solar radiatio flux data from locally stored AgERA5 data
}
\section{Functions}{
\itemize{
\item \code{get_srf.period}: Get solar radiation flux data for one location for a provided time period

\item \code{get_srf.dataset}: Iterates across a data set to extract all required data points
}}

