% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_srf.R
\name{get_srf_dp}
\alias{get_srf_dp}
\title{Extracts solar radiation flux data for one location and one date}
\usage{
get_srf_dp(.date, .lon, .lat, .agera5_folder)
}
\arguments{
\item{.date}{a date or character representing the date of the point data to be extracted}

\item{.lon}{character longitude}

\item{.lat}{character latitude}

\item{.agera5_folder}{character The location the of the folder with agera5 (*.nc) files}
}
\description{
Extracts solar radiation flux data for one location and one date
}
