% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_srf.R
\name{get_srf_ds}
\alias{get_srf_ds}
\title{Iterates across a dataset to extract all required data points}
\usage{
get_srf_ds(
  .trial_dataset,
  .start_date = "pdate",
  .end_date = "hdate",
  .lon = "lon",
  .lat = "lat",
  .agera5_folder
)
}
\arguments{
\item{.lon}{character longitude}

\item{.lat}{character latitude}

\item{.agera5_folder}{character The location the of the folder with agera5 (*.nc) files}
}
\description{
Iterates across a dataset to extract all required data points
}
