% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_srf.R
\name{get_srf_ts}
\alias{get_srf_ts}
\title{Extracts solar radiation flux data for one location for a time series}
\usage{
get_srf_ts(.start_date, .end_date, .lon, .lat, .agera5_folder)
}
\arguments{
\item{.lon}{character longitude}

\item{.lat}{character latitude}

\item{.agera5_folder}{character The location the of the folder with agera5 (*.nc) files}
}
\description{
Extracts solar radiation flux data for one location for a time series
}
