% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temp.R
\name{get_temp}
\alias{get_temp}
\alias{agera5_temp_extract}
\title{This set of functions extract temperature data from locally stored AgERA5 data}
\usage{
agera5_temp_extract(.date, .location_xy, .var, .statistic, .agera5_folder)
}
\arguments{
\item{.date}{a date or character representing the date of the point data to be extracted}

\item{.location_xy}{a data.frame or an object to be coerced, with longitud and latitude}

\item{.var}{a character of the variable of interest}

\item{.statistic}{an integer for the statistic to extract, options are:
1 = Max-24h,
2 = Max-Day-Time
3 = Mean-24h
4 = Mean-Day-Time
5 = Mean-Night-Time
6 = Min-24h
7 = Min-Night-Time}

\item{.agera5_folder}{a character with folder location}
}
\description{
This set of functions extract temperature data from locally stored AgERA5 data
}
