% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temp.R
\name{get_temp_dp}
\alias{get_temp_dp}
\title{Extracts 2m_temperature values for one location and one date}
\usage{
get_temp_dp(.date, .lon, .lat, .statistic, .agera5_folder)
}
\arguments{
\item{.date}{Date or character representing the date of the point data to be extracted}

\item{.lon}{numeric with longitude and latitude}

\item{.lat}{numeric data.frame or an object to be coerced, with longitude and latitude}

\item{.statistic}{character of the .statistic of interest, see details}

\item{.agera5_folder}{character of the .statistic of interest, see details

#'@details
\strong{valid .statistic values}
\itemize{
\item Max-24h
\item Mean-24h
\item 24_hour_minimum
\item Max-Day-Time
\item Mean-Day-Time
\item Mean-Night-Time
\item Min-Night-Time}}
}
\description{
Extracts 2m_temperature values for one location and one date
}
