% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temp.R
\name{get_temp_ds}
\alias{get_temp_ds}
\title{Iterates across a data set to extract all required data points}
\usage{
get_temp_ds(
  .trial_dataset,
  .statistic,
  .start_date = "pdate",
  .end_date = "hdate",
  .lon = "lon",
  .lat = "lat",
  .agera5_folder
)
}
\arguments{
\item{.trial_dataset}{\code{data.frame} containing the required variables, usually from the trial data points}

\item{.statistic}{character Name of the statistic to extract, see details for valid options}

\item{.start_date}{character Name of the column that holds the start date of the time period to extract}

\item{.end_date}{character Name of the column that holds the end date of the time period to extract}

\item{.lon}{character Name of the column that holds the longitude}

\item{.lat}{character Name of the column that holds the latitude}

\item{.agera5_folder}{character Location of data folder that contains AgERA5 files}
}
\value{
a list of numeric vectors with temperature values already converted from kelvin to Celsius
}
\description{
Iterates across a data set to extract all required data points
}
