% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temp.R
\name{get_temp_ts}
\alias{get_temp_ts}
\title{Extracts 2m_temperature values for one location for a time series}
\usage{
get_temp_ts(.start_date, .end_date, .lon, .lat, .statistic, .agera5_folder)
}
\arguments{
\item{.start_date}{Date or character representing the date of the point data to be extracted}

\item{.lon}{numeric with longitude and latitude}

\item{.lat}{numeric data.frame or an object to be coerced, with longitude and latitude}

\item{.statistic}{character of the .statistic of interest, see details}

\item{.agera5_folder}{character of the .statistic of interest, see details}
}
\description{
Extracts 2m_temperature values for one location for a time series
}
