% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agera5_extract.R
\name{set_agera5_source}
\alias{set_agera5_source}
\title{agera5_extract}
\usage{
set_agera5_source(.folder_path)
}
\arguments{
\item{agera5_folder}{a string with the path for the local data of AgERA5}

\item{.date}{Format dd-mm-YYYY}

\item{.location_xy}{Data frame with columns (lon, lat)}

\item{.var}{temp
prec
rel_hum}
}
\description{
Extracts data from locally stored AgERA5 data for a given location
and time period.
}
\details{
Extract data from locally stored AgERA5 data

Available temperature statistics
Max-24h
Max-Day-Time
Mean-24h
Mean-Day-Time
Mean-Night-Time
Min-24h
Min-Night-Time
}
\references{
https://cds.climate.copernicus.eu/cdsapp#!/dataset/sis-agrometeorological-indicators?tab=overview

set AgERA5 data folder
example "F:/environmental_data/AgERA5"
}
