% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{ag5_extract}
\alias{ag5_extract}
\alias{ag5_extract.default}
\alias{ag5_extract.data.frame}
\title{Extract AgERA5 data stored in a local hardrive}
\usage{
ag5_extract(variable, ..., path)

\method{ag5_extract}{default}(coords, dates, variable, ..., path)

\method{ag5_extract}{data.frame}(
  dataset,
  lon = "lon",
  lat = "lat",
  start_date = "start_date",
  end_date = "end_date",
  variable,
  ...,
  path
)
}
\arguments{
\item{variable}{The AgERA5 variable to extract, see details for available options}

\item{...}{Other parameters}

\item{path}{The path for the folder containing the AgERA5 files}

\item{coords}{numeric Vector of length = 2 of the form (lon, lat)}

\item{dates}{The dates for extracting the specified variable, either single character or a vector of length 2
or the column name in the case of \code{data.frame}}

\item{dataset}{data.frame with columns lon, lat, start_date and end_date}

\item{lon}{character Column name of longitude values in the case of \code{data.frame}}

\item{lat}{character Column name of latitude values in the case of \code{data.frame}}

\item{start_date}{Column name of start_date values in the case of \code{data.frame}}

\item{end_date}{Column name of end_date values in the case of \code{data.frame}}

\item{statistic}{character Only for variable Temperature-Air-2m, see details for valid options}

\item{time}{Only for variable Relative-Humidity-2m, see details for valid options}
}
\description{
ag5_extract is a family of wrapper functions to extract data from AgERA5 data files
previously downloaded from the Copernicus Climate Data Store. These functions use package 'terra' to
read *.nc files and extract the requested data for a given location and dates. If dates is one value
it extracts a single observation for the specified variable and location. If dates is a character vector
of \code{length == 2}, it will extract a time series of the specified variable and location, where the first
dates value is the start date and the second the end date.
}
\section{Valid variable values}{
\itemize{
\item Temperature-Air-2m
\item Precipitation-Flux
\item Solar-Radiation-Flux
\item Relative-Humidity-2m
}
}

\section{Valid statistics for variable "Temperature-Air-2m"}{
\itemize{
\item Max-24h
\item Mean-24h
\item 24_hour_minimum
\item Max-Day-Time
\item Mean-Day-Time
\item Mean-Night-Time
\item Min-Night-Time}
}

\section{Valid time values for variable "Relative-Humidity-2m"}{
\itemize{
\item 06h
\item 09h
\item 12h
\item 15h
\item 18h
}
}

\examples{
\dontrun{
2m_temp <- ag5_extract(coords = c(lon = 35.72636, lat = -2.197162),
                      dates = "1991-04-22",
                      variable = "Temperature-Air-2m",
                      statistic = "Max-Day-Time",
                      path "C:/temperature_data/")
}

}
