% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_agera5}
\alias{download_agera5}
\title{Downloads AgERA5 data from the Copernicus Climate Change Service}
\usage{
download_agera5(agera5_var, agera5_stat, day = NULL, month = NULL, year, path)
}
\arguments{
\item{agera5_var}{Character The variable to be downloaded. See details}

\item{agera5_stat}{Character Requested statistic for the requested variable. See details}

\item{day}{Character Day of the week for the requested data. NULL will download all days from requested month}

\item{month}{Character Month to be requested. NULL will download all the months for the requested year.}

\item{year}{Numeric (Integer) Year to download. Should be between 1979 - 2020}

\item{path}{Character Target location e.g. "C:/agera5" without the ending "/"}

\item{unzip}{Logical Downloaded data is provided as zip files. Should the files be uncompressed in the target path?}

\item{time}{Character}
}
\description{
The data is downloaded from Copernicus Climate Change Service (C3) using the cdsapi python library
https://github.com/ecmwf/cdsapi
This function  provides programmatic access to the dataset
For more information about the data license, please visit:
https://cds.climate.copernicus.eu/api/v2/terms/static/licence-to-use-copernicus-products.pdf
}
\details{
\strong{Allowed combinations of variable and statistic:}
\itemize{
\item cloud_cover
\itemize{
\item 24_hour_mean}
\item precipitation_flux
\item solar_radiation_flux
\item 2m_temperature
\itemize{
\item 24_hour_maximum
\item 24_hour_mean
\item 24_hour_minimum
\item day_time_maximum
\item day_time_mean
\item night_time_mean
\item night_time_minimum}
\item 2m relative humidity
\itemize{
\item 06_00
\item 09_00
\item 12_00
\item 15_00
\item 18_00}
}
}
\examples{
download_agera5(agera5_var = '2m_temperature',
agera5_stat = 'night_time_minimum',
day = NULL,
month = NULL,
year = 2015,
path = "C:/custom_target_folder/"
)
}
