% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agera5_extract.R
\name{extract_single_date}
\alias{extract_single_date}
\title{agera5_extract}
\usage{
extract_single_date(.agera5_folder, .date, .location_xy, .var, .statistic)
}
\arguments{
\item{.agera5_folder}{a string with the path for the local data of AgERA5
downloaded files should be stored per years}

\item{.date}{Format dd-mm-YYYY}

\item{.location_xy}{Data frame with columns (lon, lat)}

\item{.var}{temp
prec
rel_hum}
}
\description{
Extracts data from locally stored AgERA5 data for a given location
and time.
}
\details{
Extract data from locally stored AgERA5 data

Available temperature statistics
Max-24h
Max-Day-Time
Mean-24h
Mean-Day-Time
Mean-Night-Time
Min-24h
Min-Night-Time
}
\references{
https://cds.climate.copernicus.eu/cdsapp#!/dataset/sis-agrometeorological-indicators?tab=overview

extracts data for a given location and date
}
