% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prec.R
\name{get_prec}
\alias{get_prec}
\alias{get_prec.time_series}
\alias{get_prec.dataset}
\title{This set of functions extract precipitation data from locally stored AgERA5 data}
\usage{
get_prec(...)

\method{get_prec}{time_series}(.start_date, .end_date, .lon, .lat, .agera5_folder)

\method{get_prec}{dataset}(
  .trial_dataset = NULL,
  .start_date = "pdate",
  .end_date = "hdate",
  .lon = "lon",
  .lat = "lat",
  .agera5_folder
)
}
\arguments{
\item{.start_date}{Date or character to be coerced as Date The starting date for the period to extract}

\item{.end_date}{Date or character to be coerced as Date The end date for the period to extract}

\item{.lon}{numeric Longitude for the point of interest}

\item{.lat}{numeric Latitude for the point of interest}

\item{.agera5_folder}{Folder in the local system where the agera5 data is stored. Usually the root folder will work as the function will
search recursively}

\item{.date}{a date or character representing the date of the data point to be extracted}
}
\description{
This set of functions extract precipitation data from locally stored AgERA5 data
}
\section{Methods (by class)}{
\itemize{
\item \code{time_series}: Get precipitation data for one location for a provided time period

\item \code{dataset}: Iterates across a data set to extract all required data points
}}

