% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prec.R
\name{get_prec}
\alias{get_prec}
\alias{agera5_prec_extract}
\title{This set of functions extract precipitation data from locally stored AgERA5 data}
\usage{
agera5_prec_extract(.date, .location_xy, .agera5_folder)
}
\arguments{
\item{.date}{a date or character representing the date of the point data to be extracted}

\item{.location_xy}{a data.frame or an object to be coerced, with longitude and latitude}

\item{.var}{a character of the variable of interest}

\item{.statistic}{an integer for the statistic to extract, options are:}
}
\description{
This set of functions extract precipitation data from locally stored AgERA5 data
}
