% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temp.R
\name{get_temp}
\alias{get_temp}
\alias{get_temp.data_point}
\alias{get_temp.period}
\alias{get_temp.dataset}
\title{Extracts 2m_temperature}
\usage{
get_temp(..., .lon, .lat, .statistic, .agera5_folder)

\method{get_temp}{data_point}(.date, .lon, .lat, .statistic, .agera5_folder)

\method{get_temp}{period}(.start_date, .end_date, .lon, .lat, .statistic, .agera5_folder)

\method{get_temp}{dataset}(
  .trial_dataset = NULL,
  .stat = NULL,
  .start_date = "pdate",
  .end_date = "hdate",
  .lon = "lon",
  .lat = "lat",
  .agera5_folder
)
}
\arguments{
\item{.lon}{Character Name of the column that holds the longitude}

\item{.lat}{Character Name of the column that holds the latitude}

\item{.statistic}{an integer for the statistic to extract, see details for valid options}

\item{.agera5_folder}{a character with agera5 data folder location}

\item{.date}{Date or character representing the date of the point data to be extracted}

\item{.start_date}{Character Name of the column that holds the start date of the time period to extract}

\item{.end_date}{Character Name of the column that holds the end date of the time period to extract}

\item{.trial_dataset}{Data containing the required parameters, usually the trial data points}

\item{.var}{character of the variable of interest, currently only Temperature-Air-2m}
}
\description{
\code{get_temp} extracts temperature values of the variable 2m_temperature
from locally stored AgERA5 files
}
\details{
\strong{valid .statistic values}
\itemize{
\item Max-24h
\item Mean-24h
\item 24_hour_minimum
\item Max-Day-Time
\item Mean-Day-Time
\item Mean-Night-Time
\item Min-Night-Time}
}
\section{Methods (by class)}{
\itemize{
\item \code{data_point}: Get temperature data for one location and one date

\item \code{period}: Get temperature data for one location for a provided time period

\item \code{dataset}: Iterates across a data set to extract all required data points
}}

