% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temp.R
\name{get_temp}
\alias{get_temp}
\alias{get_temp.data_point}
\alias{get_temp.period}
\alias{get_temp.dataset}
\title{Extract temperature values from locally stored AgERA5 data}
\usage{
get_temp.data_point(.date, .lon, .lat, .var, .statistic, .agera5_folder)

get_temp.period(
  .start_date,
  .end_date,
  .lon,
  .lat,
  .var,
  .statistic,
  .agera5_folder
)

get_temp.dataset(
  .trial_dataset = NULL,
  .stat = NULL,
  .start_date = "pdate",
  .end_date = "hdate",
  .lon = "lon",
  .lat = "lat",
  .agera5_folder
)
}
\arguments{
\item{.date}{a date or character representing the date of the point data to be extracted}

\item{.lon}{Character Name of the column that holds the longitude}

\item{.lat}{Character Name of the column that holds the latitude}

\item{.var}{a character of the variable of interest}

\item{.statistic}{an integer for the statistic to extract, options are:
1 = Max-24h,
2 = Max-Day-Time
3 = Mean-24h
4 = Mean-Day-Time
5 = Mean-Night-Time
6 = Min-24h
7 = Min-Night-Time}

\item{.agera5_folder}{a character with agera5 data folder location}

\item{.start_date}{Character Name of the column that holds the start date of the time period to extract}

\item{.end_date}{Character Name of the column that holds the end date of the time period to extract}

\item{.trial_dataset}{Data containing the required parameters, usually the trial data points}

\item{.location_xy}{a data.frame or an object to be coerced, with longitude and latitude}
}
\description{
Extract temperature values from locally stored AgERA5 data
}
\details{
\strong{valid .statistic values}
\itemize{
\item 24_hour_maximum
\item 24_hour_mean
\item 24_hour_minimum
\item day_time_maximum
\item day_time_mean
\item night_time_mean
\item night_time_minimum}
}
\section{Functions}{
\itemize{
\item \code{get_temp.period}: Get temperature data for one location for a provided time period

\item \code{get_temp.dataset}: Iterates across a data set to extract all required data points
}}

